# src/swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "My Store",
      "description": "Juan Perez is a very popular guy here in my country. So I decided to create an account: id=jperez password=123456. The account has the 'guest' role which means than. Once you login with that id and password I recommend don't refresh the page. If you do, you'll have to login one more time",
      "contact": {
        "name": "Sandro Simón",
        "email": "hello@sandrosimon.com"
      },
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "/api/v1",
        "description": "Server"
      }
      
    ],
...
    "paths": {
...
      "/products": {
        "get": {
          "tags": [
            "Product"
          ],
          "summary": "User roles with access to this endpoint: admin, seller, guest",
          "description": "Get all products",
          "operationId": "getAllProducts",
          "responses": {
            "200": {
              "description": "Successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref":"#/components/schemas/Product"
                    }

                  }
                  
                }
              }
            }
          },
          "security": [
            { "storeAuth": [] }
          ]
        },
        "post": {
          "tags": [
            "Product"
          ],
          "summary": "User roles with access to this endpoint: admin",
          "description": "Add a new account to the database",
          "operationId": "createOneProduct",
          "requestBody": {
            "description": "Create a new product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref":"#/components/schemas/Product"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref":"#/components/schemas/Product"  
                  }
                }
                
              }
            },
            "405": {
              "description": "Invalid input"
            }
          },
          "security": [
            { "storeAuth": [] }
          ]
        }      
      },
      "/products/{id}": {
        "get": {
          "tags": [
            "Product"
          ],
          "summary": "User roles with access to this endpoint 👉 admin, seller, guest",
          "description": "Returns an account",
          "operationId": "getOneProduct",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "Id of the product to return",
              "schema": {
                "type": "string",
                "pattern": "^[A-Z]{1}[0-9]{3}$",
                "example": "A100"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref":"#/components/schemas/Product"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Product not found"
            }
          },
          "security": [
            { "storeAuth": [] }
          ]
        },
        "put": {
          "tags": [
            "Product"
          ],
          "summary": "User roles with access to this endpoint: admin",
          "description": "Update an existing account by Id",
          "operationId": "updateOneAccount",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Id of the product to return.",
              "required": true,
              "schema": {
                "type": "string",
                "pattern": "^[A-Z]{1}[0-9]{3}$",
                "example": "A100"
              }
            } 
          ],
          "requestBody": {
            "description": "Update an existent product in the store.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                },
                "example": {
                  "description": "LECHE AMARILLA",
	                "uom": "UND",
	                "stock": 0.00,
	                "lastPrice": 5.8,
                	"suggestedPrice": 7
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid account code supplied"
            },
            "404": {
              "description": "Account not found"
            },
            "405": {
              "description": "Validation exception"
            }
          },
          "security": [
            { "storeAuth": [] }
          ]
          
        },
        "delete": {
          "tags": [
            "Product"
          ],
          "summary": "User roles with access to this endpoint: admin",
          "description": "delete an product",
          "operationId": "deleteOneProduct",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Product id to delete",
              "required": true,
              "schema": {
                "type": "string",
                "pattern":"^[A-B]{1}[0-9]{3}$",
                "example":"A100"
              }
            }
          ],
          "responses": {
            "200":{
              "description": "The item was succesfully removed."
            },
            "500":{
              "description": "Unexpected error."
            }
          },
          "security": [
            { "storeAuth": [] }
          ]
        }
        
      },
...
      "/orders": {
        "get": {
          "tags": ["Order"],
          "summary": "User roles with access to this endpoint: admin, seller, guest",
          "description": "Get all orders",
          "operationId": "getAllOrders",
          "responses": {
            "200": {
              "description": "Successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref":"#/components/schemas/Order"
                    }  
                  }
                }
              }
            }
          },
          "security": [
            { "storeAuth": [] }
          ]
        },
        "post": {
          "tags": ["Order"],
          "summary": "User roles with access to this endpoint: admin, seller",
          "description": "Add a new transaction to the database",
          "operationId": "createOrder",
          "requestBody": {
            "description": "Create a new order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                },
                "example": {
                  "date": "2022-10-05",
                  "customerId": "22222222",
                  "items": [
                    {
                      "productId": "A100",
                      "quantity": "5.00",
                      "unitPrice": "4.50"
                    }
                  ]
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref":"#/components/schemas/Order"  
                  }
                }
                
              }
            },
            "405": {
              "description": "Invalid input"
            }
          },
          "security": [
            { "storeAuth": [] }
          ]
        }      
      },
      "/orders/{id}": {
        "get": {
          "tags": ["Order"],
          "summary": "User roles with access to this endpoint: admin, seller, guest",
          "description": "Returns a order",
          "operationId": "getOneOrder",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Id of the order to return",
              "required": true,
              "schema": {
                "type": "integer",
                "format":"int64",
                "example": 1
              }
              
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref":"#/components/schemas/Order"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Account not found"
            }
          },
          "security": [
            { "storeAuth": [] }
          ]
        },
        "put": {
          "tags": ["Order"],
          "summary": "User roles with access to this endpoint: admin, seller",
          "description": "Update an existing order by Id",
          "operationId": "updateOrder",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Code of the order to update",
              "required": true,
              "schema": {
                "type": "number",
                "format":"int64",
                "example":1
              }
            } 
          ],
          "requestBody": {
            "description": "Update an existent transaction in the database.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                },
                "example": {
                    "date": "2022-10-07",
                    "customerId": "40483785",
                    "items": [
                      {
                        "productId": "A200",
                        "quantity": "1.00",
                        "unitPrice": "50"
                      }
                    ]
                  }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid transaction Id supplied"
            },
            "404": {
              "description": "Transaction not found"
            },
            "405": {
              "description": "Validation exception"
            }
          },
          "security": [
            { "storeAuth": [] }
          ]
          
        },
        "delete": {
          "tags": ["Order"],
          "summary": "User roles with access to this endpoint: admin",
          "description": "delete a transaction",
          "operationId": "deleteOrder",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Order id to delete",
              "required": true,
              "schema": {
                "type": "integer",
                "format":"int64",
                "example": 1
              }
            }
          ],
          "responses": {
            "204":{
              "description": "The transaction was deleted successfully."
            },
            "500":{
              "description": "Unexpected error."
            }
          },
          "security": [
            { "storeAuth": [] }
          ]
        }
      },
...
    },
    "components": {
      "schemas": {
        "Product": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "example": "A100"
            },
            "description": {
              "type": "string",
              "example": "JABON"
            },
            "uom": {
              "type": "string",
              "example": "UND"
            },
            "stock": {
              "type": "number",
              "example": 10
            },
            "lastPrice": {
              "type": "number",
              "example": 25
            },
            "suggestedPrice": {
              "type": "number",
              "example": 100.00
            }
          }
        },
...
        "Order": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format":"int64"
            },
            "date": {
              "type": "string"
            },
            "customerId": {
              "type": "string"
            },
            "salespersonId": {
              "type": "string"
            },
            "items": {
              "type":"array"
            }
          }
        }
      },
      "securitySchemes": {
        "storeAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    }  
  }